feat(commonware-node): update forkchoice after propose, verify (#305)
Moves updating forkchoice state from the finalizer task to a new
forkchoice-updater task.

The finalizer task no longer sends forkchoice-updates on every
finalization, but merely tracks the last finalized block digest, making
it available through a watch channel (NOTE: breaking convention of only
using futures-rs items because tokio watch channels are more ergonomic).

Instead of the finalizer, a forkchoice-updater task is now responsible
for updating the canonical chain. It relies on the fact that
verifications and proposals are guaranteed to be on notarized parents,
and on the finalizer making its latest finalized block available.

1. at the beginning of each proposal, setting `head = proposal.parent`
and `finalized = last_finalized_hash`.
2. after a succesful proposal, setting `head = proposal` and `finalized
= last_finalized_hash`.
3. at the beginning of each verification, setting `head =
proposal.parent` and `finalized = last_finalized_hash`.
4. after a succesful verification, setting `head = proposal` and
`finalized = last_finalized_hash`.

We choose to set forkchoice state before and after a proposal to defend
against block building taking very long: if we wait until after a block
was successfully built to decide whether to set the head to parent or
proposal, we can potentially stall a node from progressing for several
seconds.

For verification we similarly know that we always want to set the
forkchoice state to at least the notarized parent - and, if successful,
to the verified proposal derived from it. With forkchoice running in the
background there is no need to stall chain evolution on verification
running through.