fix: add revert if policy doesn't exist (#1801)
Addresses TMPO 29


It reverts when a policy doesn't exist, instead of returning false. This
removes ambiguity from errors, for what is causing the issue.

It does incur one extra storage read of the policyId counter per policy
fetch. I don't see a way to avoid this. Maybe the perf degradation is
not worth the fix, because it primarily affects offchain tooling, and it
should be pretty easy to check offchain if a policy exists or not.

CC: @0xKitsune